requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.1.0
openpyxl>=3.1.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0
click>=8.1.0
rich>=13.0.0
supabase>=2.0.0
flask>=3.0.0
gunicorn>=21.2.0